config: Fix crash in reload if new interfaces are added
This is a bug I seem to have introduced in
429209f4aa3c55504a49833e0004489f241e7819 where we compare links
for changes. if a new node was added on an existing link then it
was compared against a non-existant one in the previous configuration.
We now only compare nodes that are in both interfaces.

As I needed min() for this function, I moved it from individual
.c files into util.h so we only have one copy.

And the error message was fixed.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>