pkgconfig: Add required private libraries
Most of the libraries requires libcorosync_common and libqb. These
libraries are using modified libtemplate.pc.in as before.

Exception is libcorosync_common (with no requirements) and libsam
(which requires libcmap and libquorum) so these have to have specific
files and make rules.

Also corosync.pc.in is removed because it looks like leftover from
old plugins times.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>