totemknet: Add support for libnozzle devices
A nozzle device is a pseudo ethernet device that routes network
traffic through a channel on the corosync knet network (NOT cpg or any
corosync  internal  service) to other nodes in the cluster. It allows
applications to take advantage of knet features such  as multipathing,
automatic failover, link switching etc.