corosync-quorumtool: print totem cluster information
print some common cluster information from the cmap totem section.

For now print cluster name, config version, transport method and
secauth. This is information which is often relevant when wanting to
checkout a summary of the cluster state.

The fallback to default is hardcoded, which is something to be
improved, else the implementation is quite simple. An alternative
could be to re-use parts of the print_key method from
corosync-cmapctl, which would save us some hassle here and would make
this more general.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>