[stats] fixes for prefix tracking.
Let me know if I've missed something here.

Tracking of "" is now marked as NOT_SUPPORTED when it's attempted,
which makes far more sense than storing it and then ignoring it every
time.

Prefix tracking of schedmiss is allowed if you really want to
know when the 0..9 keys are created. I suppose it could be handy after a
clear operation to know if events are getting common (though a simple
modify tracker on .0. is probably as good tbh, but I'm not really in
favour of hiding information for no good reason).

I've fixed the order of clearing the values so that a delete tracker
will get the correct values.