sam: Refactor locking
Locking in sam was simply completely broken so this rather huge patch
tries to fix it.

Basically all public functions are now locked before any access
of sam_internal_data is made. This works with exception of
public function which are also called by parent, so nolock of sam_stop,
sam_store_data and sam_warn_signal_set is added and called by parent.

Also size of stored data is checked to never be equal or larger to
maximum ssize_t (so basically it is artificial check).

Signed-off-by: Jan Friesse <jfriesse@redhat.com>
Reviewed-by: Christine Caulfield <ccaulfie@redhat.com>