sam: Do not lock sam_register
sam_register is function which is really not intended to be
ever called in any other thread than main one so it is safe to not use
the lock to prevent possible problems with fork.

Special thanks to Christine Caulfield <ccaulfie@redhat.com> for
mentioning possible problems with mutex and fork.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>